Conflicts:
	include/Flashcard.h
	include/Hint.h
	include/SmartPicker.h
	include/WordData.h
	include/wordData.h
	src/Flashcard.cpp
	src/Hint.cpp
	src/LanguageTest.cpp
	src/SmartPicker.cpp
	src/WordData.cpp
	src/functions.cpp
	src/wordData.cpp

+++ include/WordData.h
@@ -16,13 +16,14 @@ using std::cout;
 using std::endl;
 
 class WordData
+
 {
- public:
+public:
   WordData();
   unsigned int numAsked;
   unsigned int numCorrect;
-  double answeringTime;
-  double getPercentCorrect();
-  double getAverageTime();
+  float answeringTime;
+  float getPercentCorrect();
+  float getAverageTime();
 };
 #endif // _WORDDATA_H_
